Working standalone installation as a console_script
